feat: enhance DeskSearch and Screen components; streamline search history handling and improve UI interactions
